// server/src/api/types.ts
export interface AuthRequired {
  Authorization: string;
}
// ...
export interface ListUsersRequest {
  filter?: ListUsersFilter;
  maxPerPage?: number;
  page?: number;
}

export interface ListUsersFilter {
  role?: Role;
  name?: string;
}

export interface ListUsersResponse {
  results: UserDetails[];
  totalResults: number;
  pageCount: number;
}

export interface UserDetails {
  userId: string;
  email: string;
  name: string;
  role: Role;
}

export type Role = "client" | "realtor" | "admin";
// ...
export type ListUsers_Response = {
  kind: "success";
  data: ListUsersResponse;
} | {
  kind: "unauthorized";
  data: string;
};